backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arize-otel>=0.12.0",
    "chromadb==1.5.5",
    "docx2txt==0.8",
    "fastapi==0.115.5",
    "google-generativeai==0.8.3",
    "langchain==1.2.13",
    "langchain-chroma==1.1.0",
    "langchain-classic==1.0.3",
    "langchain-community==0.4.1",
    "langchain-google-genai==4.2.1",
    "langchain-text-splitters>=1.1.1,<2.0.0",
    "openinference-instrumentation-langchain>=0.1.61",
    "opentelemetry-exporter-otlp>=1.40.0",
    "opentelemetry-sdk>=1.40.0",
    "pydantic==2.10.3",
    "pypdf==6.9.2",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.17",
    "uvicorn[standard]==0.32.1",
]
